+++ src/plugin.hpp
@@ -6,16 +6,7 @@ using namespace rack;
 
 extern Plugin *pluginInstance;
 
-// Declare each Model, defined in each module source file
-// extern Model* modelMyModule;
-extern Model* modelBss;
-extern Model* modelCrcl;
-extern Model* modelNtrvlc;
-extern Model* modelNtrvlx;
-extern Model* modelSpc;
-extern Model* modelStpr;
-extern Model* modelTxt;
-extern Model* modelWhl;
+extern Model *modelBrst;
 extern Model *modelBss;
 extern Model *modelCrcl;
 extern Model *modelNtrvlc;
@@ -23,4 +14,5 @@ extern Model *modelNtrvlx;
 extern Model *modelSpc;
 extern Model *modelStpr;
 extern Model *modelTxt;
+extern Model *modelVbrt;
 extern Model *modelWhl;